feat: enable :chain/detail/:id prefix
